requirements.txt
requests
pyyaml
python-slugify
bs4
